pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flexible-rep-counter"
version = "0.1.0"
description = "Flexible rep counter engine (landmarks -> reps) with optional local OpenCV visualizer app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
viz = [
    "opencv-python>=4.8.0",
    "requests>=2.31.0",
]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
